Revert touching routes code
